webgl/missing_arguments.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>WebGL All Of Interface Argument Errors.</title>
  <link rel='author' href='mailto:kawada.hirosi@gmail.com" title='kawada.hirosi'>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>missing_arguments</h1>
    <div id="log"></div>
	<canvas id="testcanvas"></canvas>
    <script>
		var canvas = document.getElementById("testcanvas");
		var webgl = canvas.getContext("webgl") || canvas.getContext("experimental-webgl");
        var tests = [
            function() { webgl.activeTexture(); },
            function() { webgl.bindBuffer(); },
            function() { webgl.bindFramebuffer(); },
            function() { webgl.bindRenderbuffer(); },
            function() { webgl.blendColor( 0.0, 0.0, 0.0 ); },
            function() { webgl.blendColor( 0.0, 0.0 ); },
            function() { webgl.blendColor( 0.0 ); },
            function() { webgl.blendColor(); },
            function() { webgl.blendEquation(); },
            function() { webgl.blendEquationSeparate( 0.0 ); },
            function() { webgl.blendEquationSeparate(); },
            function() { webgl.blendFunc( null ); },
            function() { webgl.blendFunc(); },
            function() { webgl.blendFuncSeparate( null, null, null ); },
            function() { webgl.blendFuncSeparate( null, null ); },
            function() { webgl.blendFuncSeparate( null ); },
            function() { webgl.blendFuncSeparate(); },
            function() { webgl.bufferData( null, null ); },
            function() { webgl.bufferData(); },
            function() { webgl.checkFramebufferStatus(); },
            function() { webgl.clear(); },
            function() { webgl.clearColor( 0.0, 0.0, 0.0 ); },
            function() { webgl.clearColor( 0.0, 0.0 ); },
            function() { webgl.clearColor( 0.0 ); },
            function() { webgl.clearDepth(); },
            function() { webgl.clearStencil(); },
            function() { webgl.colorMask( 0.0, 0.0, 0.0 ); },
            function() { webgl.colorMask( 0.0, 0.0 ); },
            function() { webgl.colorMask( 0.0 ); },
            function() { webgl.colorMask(); },
            function() { webgl.compressedTexImage2D( null, 0, 0, 0, 0, 0 ); },
            function() { webgl.compressedTexImage2D( null, 0, 0, 0, 0 ); },
            function() { webgl.compressedTexImage2D( null, 0, 0, 0 ); },
            function() { webgl.compressedTexImage2D( null, 0, 0 ); },
            function() { webgl.compressedTexImage2D( null, 0 ); },
            function() { webgl.compressedTexImage2D( null ); },
            function() { webgl.compressedTexImage2D(); },
            function() { webgl.compressedTexSubImage2D( null, null, null, null, null, null, null ); },
            function() { webgl.compressedTexSubImage2D( null, null, null, null, null  ); },
            function() { webgl.compressedTexSubImage2D( null, null, null, null, null ); },
            function() { webgl.compressedTexSubImage2D( null, null, null, null  ); },
            function() { webgl.compressedTexSubImage2D( null, null, null ); },
            function() { webgl.compressedTexSubImage2D( null, null ); },
            function() { webgl.compressedTexSubImage2D( null ); },
            function() { webgl.compressedTexSubImage2D(); },
            function() { webgl.copyTexImage2D( null, null, null, null, null, null, null ); },
            function() { webgl.copyTexImage2D( null, null, null, null, null, null ); },
            function() { webgl.copyTexImage2D( null, null, null, null, null ); },
            function() { webgl.copyTexImage2D( null, null, null, null ); },
            function() { webgl.copyTexImage2D( null, null, null ); },
            function() { webgl.copyTexImage2D( null, null ); },
            function() { webgl.copyTexImage2D( null ); },
            function() { webgl.copyTexImage2D(); },
            function() { webgl.copyTexSubImage2D( null, null, null, null, null, null, null ); },
            function() { webgl.copyTexSubImage2D( null, null, null, null, null, null ); },
            function() { webgl.copyTexSubImage2D( null, null, null, null, null ); },
            function() { webgl.copyTexSubImage2D( null, null, null, null ); },
            function() { webgl.copyTexSubImage2D( null, null, null ); },
            function() { webgl.copyTexSubImage2D( null, null ); },
            function() { webgl.copyTexSubImage2D( null ); },
            function() { webgl.copyTexSubImage2D(); },
            function() { webgl.createShader(); },
            function() { webgl.cullFace(); },
            function() { webgl.depthFunc(); },
            function() { webgl.depthMask(); },
            function() { webgl.depthRange( null ); },
            function() { webgl.disable(); },
            function() { webgl.disableVertexAttribArray(); },
            function() { webgl.drawArrays( null, null ); },
            function() { webgl.drawArrays( null ); },
            function() { webgl.drawArrays(); },
            function() { webgl.drawElements( null, null, null ); },
            function() { webgl.drawElements( null, null ); },
            function() { webgl.drawElements( null ); },
            function() { webgl.drawElements(); },
            function() { webgl.enable(); },
            function() { webgl.enableVertexAttribArray(); },
            function() { webgl.framebufferRenderbuffer( null, null ); },
            function() { webgl.framebufferRenderbuffer( null ); },
            function() { webgl.framebufferRenderbuffer(); },
            function() { webgl.framebufferTexture2D( null, null, null, null ); },
            function() { webgl.framebufferTexture2D( null, null ); },
            function() { webgl.framebufferTexture2D( null ); },
            function() { webgl.framebufferTexture2D(); },
            function() { webgl.frontFace(); },
            function() { webgl.generateMipmap(); },
            function() { webgl.getActiveAttrib( null ); },
            function() { webgl.getActiveAttrib(); },
            function() { webgl.getActiveUniform( null ); },
            function() { webgl.getActiveUniform(); },
            function() { webgl.getAttribLocation( null ); },
            function() { webgl.getAttribLocation(); },
            function() { webgl.getBufferParameter( null ); },
            function() { webgl.getBufferParameter(); },
            function() { webgl.getParameter(); },
            function() { webgl.getFramebufferAttachmentParameter( null, null ); },
            function() { webgl.getFramebufferAttachmentParameter( null ); },
            function() { webgl.getFramebufferAttachmentParameter(); },
            function() { webgl.getProgramParameter( null, null ); },
            function() { webgl.getProgramParameter(); },
            function() { webgl.getRenderbufferParameter( null ); },
            function() { webgl.getRenderbufferParameter(); },
            function() { webgl.getShaderParameter( null ); },
            function() { webgl.getShaderPrecisionFormat( null ); },
            function() { webgl.getShaderPrecisionFormat(); },
            function() { webgl.getTexParameter( null, null ); },
            function() { webgl.getTexParameter( null ); },
            function() { webgl.getTexParameter(); },
            function() { webgl.getUniformLocation( null ); },
            function() { webgl.getUniformLocation(); },
            function() { webgl.getVertexAttrib( null ); },
            function() { webgl.getVertexAttrib(); },
            function() { webgl.getVertexAttribOffset( null ); },
            function() { webgl.getVertexAttribOffset(); },
            function() { webgl.hint( null ); },
            function() { webgl.hint(); },
            function() { webgl.isEnabled(); },
            function() { webgl.lineWidth(); },
            function() { webgl.pixelStorei( null ); },
            function() { webgl.pixelStorei(); },
            function() { webgl.polygonOffset( null ); },
            function() { webgl.polygonOffset(); },
            function() { webgl.readPixels( null, null, null, null ); },
            function() { webgl.readPixels( null, null, null ); },
            function() { webgl.readPixels( null, null ); },
            function() { webgl.readPixels( null ); },
            function() { webgl.readPixels(); },
            function() { webgl.renderbufferStorage( null, null, null ); },
            function() { webgl.renderbufferStorage( null, null ); },
            function() { webgl.renderbufferStorage( null ); },
            function() { webgl.renderbufferStorage(); },
            function() { webgl.sampleCoverage( null ); },
            function() { webgl.sampleCoverage(); },
            function() { webgl.scissor( null, null, null ); },
            function() { webgl.scissor( null, null ); },
            function() { webgl.scissor( null ); },
            function() { webgl.scissor(); },
            function() { webgl.shaderSource(); },
            function() { webgl.stencilFunc( null, null ); },
            function() { webgl.stencilFunc( null ); },
            function() { webgl.stencilFunc(); },
            function() { webgl.stencilFuncSeparate( null, null, null ); },
            function() { webgl.stencilFuncSeparate( null, null ); },
            function() { webgl.stencilFuncSeparate( null ); },
            function() { webgl.stencilFuncSeparate(); },
            function() { webgl.stencilMask(); },
            function() { webgl.stencilMaskSeparate( null ); },
            function() { webgl.stencilMaskSeparate(); },
            function() { webgl.stencilOp( null, null ); },
            function() { webgl.stencilOp( null ); },
            function() { webgl.stencilOp(); },
            function() { webgl.stencilOpSeparate( null, null, null ); },
            function() { webgl.stencilOpSeparate( null, null ); },
            function() { webgl.stencilOpSeparate( null ); },
            function() { webgl.stencilOpSeparate(); },
            function() { webgl.texImage2D( null, null, null, null, null, null, null ); },
            function() { webgl.texImage2D( null, null, null, null, null, null ); },
            function() { webgl.texImage2D( null, null, null, null, null ); },
            function() { webgl.texImage2D( null, null, null, null ); },
            function() { webgl.texImage2D( null, null, null ); },
            function() { webgl.texImage2D( null, null ); },
            function() { webgl.texImage2D( null ); },
            function() { webgl.texImage2D(); },
            function() { webgl.texParameterf( null, null ); },
            function() { webgl.texParameterf( null ); },
            function() { webgl.texParameterf(); },
            function() { webgl.texSubImage2D( null, null, null, null, null, null, null ); },
            function() { webgl.texSubImage2D( null, null, null, null, null, null ); },
            function() { webgl.texSubImage2D( null, null, null, null, null ); },
            function() { webgl.texSubImage2D( null, null, null, null ); },
            function() { webgl.texSubImage2D( null, null, null ); },
            function() { webgl.texSubImage2D( null, null ); },
            function() { webgl.texSubImage2D( null ); },
            function() { webgl.texSubImage2D(); },
            function() { webgl.uniform1f( null ); },
            function() { webgl.uniform1f(); },
            function() { webgl.uniform1fv( null ); },
            function() { webgl.uniform1fv(); },
            function() { webgl.uniform1i( null ); },
            function() { webgl.uniform1i(); },
            function() { webgl.uniform1iv( null ); },
            function() { webgl.uniform1iv(); },
            function() { webgl.uniform2f( null, null ); },
            function() { webgl.uniform2f( null ); },
            function() { webgl.uniform2f(); },
            function() { webgl.uniform2fv( null ); },
            function() { webgl.uniform2fv(); },
            function() { webgl.uniform2iv( null ); },
            function() { webgl.uniform2iv(); },
            function() { webgl.uniform3f(); },
            function() { webgl.uniform3fv( null ); },
            function() { webgl.uniform3fv(); },
            function() { webgl.uniform3iv( null ); },
            function() { webgl.uniform3iv(); },
            function() { webgl.uniform4f(); },
            function() { webgl.uniform4fv( null ); },
            function() { webgl.uniform4fv(); },
            function() { webgl.uniform4iv( null ); },
            function() { webgl.uniform4iv(); },
            function() { webgl.uniformMatrix2fv( null, null ); },
            function() { webgl.uniformMatrix2fv( null ); },
            function() { webgl.uniformMatrix2fv(); },
            function() { webgl.uniformMatrix3fv( null, null ); },
            function() { webgl.uniformMatrix3fv( null ); },
            function() { webgl.uniformMatrix3fv(); },
            function() { webgl.uniformMatrix4fv( null, null ); },
            function() { webgl.uniformMatrix4fv( null ); },
            function() { webgl.uniformMatrix4fv(); },
            function() { webgl.vertexAttrib1f( null ); },
            function() { webgl.vertexAttrib1f(); },
            function() { webgl.vertexAttrib1fv( null ); },
            function() { webgl.vertexAttrib1fv(); },
            function() { webgl.vertexAttrib2f( null ); },
            function() { webgl.vertexAttrib2f(); },
            function() { webgl.vertexAttrib2fv( null ); },
            function() { webgl.vertexAttrib2fv(); },
            function() { webgl.vertexAttrib3f( null ); },
            function() { webgl.vertexAttrib3f(); },
            function() { webgl.vertexAttrib3fv( null ); },
            function() { webgl.vertexAttrib3fv(); },
            function() { webgl.vertexAttrib4f( null ); },
            function() { webgl.vertexAttrib4f(); },
            function() { webgl.vertexAttrib4fv( null ); },
            function() { webgl.vertexAttribPointer( null, null, null, null, null ); },
            function() { webgl.vertexAttribPointer( null, null, null, null ); },
            function() { webgl.vertexAttribPointer( null, null, null ); },
            function() { webgl.vertexAttribPointer( null, null ); },
            function() { webgl.vertexAttribPointer( null ); },
            function() { webgl.vertexAttribPointer(); },
            function() { webgl.viewport( null, null, null ); },
            function() { webgl.viewport( null, null ); },
            function() { webgl.viewport( null ); },
            function() { webgl.viewport(); },
        ];
        tests.forEach(function(fun) {
            test(function() {
                assert_throws(new TypeError(), fun);
            } , "Should throw TypeError for " + fun.toString() +".");
        }); // Web IDL spec (http://dev.w3.org/2006/webapi/WebIDL/#es-interface-call)
    </script>
 </body>
</html>
